Factor out pymalloc.h.
